SRE-326: Update Collector config for Honeycomb E&S migration
We'll be double-sending all our data for 30 days. This requires a
second, separate API key. I've added a new ExternalSecretPush and
updated the deployment to load it into a new environment variable.

For trace data, we only need to add a second exporter to send to the new
environment and add that exporter to the pipeline.

For metrics data, Honeycomb doesn't automatically break it out into
service datasets. So even though the prometheus `job_name` corresponds
to `service.name` in Honeycomb, we won't have to change our current
configuration. The update here is the same as for trace data: just
adding a second exporter to send to the new environment.